+++ src/services/ui/src/components/SearchForm/index.tsx
@@ -1,0 +1,46 @@
+export const SearchForm = ({
+  handleSearch,
+  searchText,
+  setSearchText,
+}: {
+  handleSearch: (searchString: string) => Promise<void>;
+  searchText: string;
+  setSearchText: React.Dispatch<React.SetStateAction<string>>;
+}) => {
+  const handleSubmit = (event: React.FormEvent<HTMLFormElement>) => {
+    event.preventDefault();
+    handleSearch(searchText);
+  };
+
+  const handleInputChange = (event: React.ChangeEvent<HTMLInputElement>) => {
+    setSearchText(event.target.value);
+  };
+
+  return (
+    <form onSubmit={handleSubmit}>
+      <div className="relative">
+        <svg
+          xmlns="http://www.w3.org/2000/svg"
+          className="absolute top-0 bottom-0 w-6 h-6 my-auto text-gray-400 left-3"
+          fill="none"
+          viewBox="0 0 24 24"
+          stroke="currentColor"
+        >
+          <path
+            strokeLinecap="round"
+            strokeLinejoin="round"
+            strokeWidth={2}
+            d="M21 21l-6-6m2-5a7 7 0 11-14 0 7 7 0 0114 0z"
+          />
+        </svg>
+        <input
+          type="text"
+          placeholder="Search"
+          className="w-full py-3 pl-12 pr-4 text-gray-500 border border-gray-300 outline-none focus:bg-white focus:border-indigo-600"
+          value={searchText}
+          onChange={handleInputChange}
+        />
+      </div>
+    </form>
+  );
+};
