+++ src/services/ui/src/components/index.tsx
@@ -2,3 +2,4 @@ export * from "./LoadingSpinner";
 export * from "./ErrorAlert";
 export * from "./Layout";
 export * from "./HowItWorks";
+export * from "./SearchForm";
